improve coverage

+++ pytest_play/providers/python.py
@@ -68,7 +68,7 @@ class PythonProvider(BaseProvider):
     def command_sleep(self, command, **kwargs):
         """ Exec and return an expression
         """
-        wait_time = int(command['seconds'])
+        wait_time = float(command['seconds'])
         sleep(wait_time)
 
     def command_wait_until(self, command, **kwargs):
